Updated to BigIntField for large values

+++ discordoauth/models.py
@@ -6,8 +6,8 @@ class DiscordOAuth(models.Model):
     user = models.ForeignKey("schoolboxauth.User", on_delete=models.CASCADE)
     access_token = models.CharField(max_length=64)
     refresh_token = models.CharField(max_length=64)
-    discord_id = models.IntegerField(null=True)
-    expires = models.IntegerField()
+    discord_id = models.BigIntegerField(null=True)
+    expires = models.BigIntegerField()
 
     def __str__(self):
         return self.user.name
